# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "Magalu Produtos API",
    "description": "Api para consumir endpoint relacionados aos produtos da Magalu."
  },
  "host": "localhost:3003",
  "basePath": "/",
  "tags": [
    {
      "name": "Produtos",
      "description": "Endpoints relacionados aos recursos de produto."
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Insira seu token aqui."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/produtos/": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para obter uma lista de produtos.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "schema": {
              "$ref": "#/definitions/Produto"
            }
          },
          "400": {
            "description": "Houve algum problema na requisição."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Lista não encontrada."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para criar um novo produto.",
        "parameters": [
          {
            "name": "novoProduto",
            "in": "body",
            "description": "Informações do produto",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/NovoProduto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Produto criado"
          },
          "400": {
            "description": "Os campos nome e preço são obrigatórios."
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/fornecedores/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/usuarios/": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/usuarios/auth": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Produto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nome": {
          "type": "string",
          "example": "DELL"
        },
        "preco": {
          "type": "number",
          "example": 5000
        },
        "create_at": {
          "type": "string",
          "example": "2021-01-09 13:05:36"
        },
        "updated_at": {
          "type": "string",
          "example": "2021-01-09 13:05:40"
        }
      }
    },
    "NovoProduto": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Dell"
        },
        "preco": {
          "type": "number",
          "example": 5000
        }
      },
      "required": [
        "nome",
        "preco"
      ]
    }
  }
}